feat: 판매 집계 스케줄러 / 파사드 수정
